feat(file): 업로드시, 파일 변조 체크(확장자)

+++ src/main/java/com/github/dankook_univ/meetwork/file/application/file/FileServiceImpl.java
@@ -9,6 +9,7 @@ import com.github.dankook_univ.meetwork.file.exceptions.NotSupportedFileFormatEx
 import com.github.dankook_univ.meetwork.file.infra.persistence.FileRepositoryImpl;
 import com.github.dankook_univ.meetwork.member.application.MemberServiceImpl;
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
@@ -23,6 +24,8 @@ import org.springframework.web.multipart.MultipartFile;
 public class FileServiceImpl implements FileService {
 
     private static final Tika tika = new Tika();
+    private static final List<String> FILETYPE = Arrays.asList("image/jpeg", "image/png",
+        "image/jpg");
     private final FileRepositoryImpl fileRepository;
     private final StorageServiceImpl storageService;
     private final MemberServiceImpl memberService;
@@ -30,47 +33,38 @@ public class FileServiceImpl implements FileService {
     @Override
     @Transactional
     public File upload(String memberId, FileType fileType, MultipartFile multipartFile) {
+        File file = null;
         try {
-            String tikaChecked = tika.detect(multipartFile.getInputStream());
-            System.out.println("checked mime : " + tikaChecked);
+            String mime = tika.detect(multipartFile.getInputStream());
+            if (!FILETYPE.contains(mime)) {
+                throw new NotSupportedFileFormatException(mime);
+            }
+            mime = mime.substring(mime.lastIndexOf("/") + 1).toLowerCase();
+
+            file = fileRepository.save(
+                File.builder()
+                    .uploader(memberService.getById(memberId))
+                    .type(fileType)
+                    .mime(mime)
+                    .name(
+                        Objects.requireNonNull(multipartFile.getOriginalFilename())
+                            .replaceAll("\\.", "")
+                            .replaceAll("/", "")
+                            .replaceAll("\\\\", "")
+                    )
+                    .build()
+            );
+
+            storageService.upload(multipartFile, file.getKey());
+
         } catch (IOException e) {
             throw new FailedToFileUploadException();
-        }
-
-        String mime = multipartFile.getOriginalFilename() == null
-            ? "jpeg"
-            : multipartFile.getOriginalFilename().substring(
-                multipartFile.getOriginalFilename().lastIndexOf(".") + 1
-            ).toLowerCase();
-        if (
-            !"gif".equals(mime)
-                && !"jpg".equals(mime)
-                && !"png".equals(mime)
-                && !"jpeg".equals(mime)
-        ) {
-            throw new NotSupportedFileFormatException(mime);
-        }
-
-        File file = fileRepository.save(
-            File.builder()
-                .uploader(memberService.getById(memberId))
-                .type(fileType)
-                .mime(mime)
-                .name(
-                    Objects.requireNonNull(multipartFile.getOriginalFilename())
-                        .replaceAll("\\.", "")
-                        .replaceAll("/", "")
-                        .replaceAll("\\\\", "")
-                )
-                .build()
-        );
-
-        try {
-            storageService.upload(multipartFile, file.getKey());
-            return file;
         } catch (FailedToFileUploadException e) {
-            fileRepository.delete(file.getId().toString());
-            return null;
+            if (file != null) {
+                fileRepository.delete(file.getId().toString());
+            }
+        } finally {
+            return file;
         }
     }
